Added OCR_GroundTruth.ipynb
This is the code to run OCR to scan through all our dataset to generate ground truth. Note that the code is incomplete as the methodology to save the json ground truth has not been determined. The overall OCR already works but parsing is not completely accurate, that is as expected.